refactor(merkle_tree): lift VersionedMerkleTree into a lib crate
Move VersionedMerkleTree and its tree-output adapter out of node/bin
into the new versioned_merkle_tree crate. The ZiSK witness crate needs
this type, so a shared crate lets both the server and the witness crate
use it. Move the related merkle-query metrics into the new crate too.
This change moves code only. The behavior does not change.
