+++ lib/versioned_merkle_tree/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "versioned_merkle_tree"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+zksync_os_merkle_tree.workspace = true
+
+zk_ee.workspace = true
+zk_os_basic_system.workspace = true
+zk_os_forward_system.workspace = true
+
+alloy = { workspace = true, default-features = false }
+anyhow.workspace = true
+tracing.workspace = true
+vise.workspace = true
