+++ node/bin/Cargo.toml
@@ -60,6 +60,7 @@ zksync_os_replay_archive.workspace = true
 zksync_os_state.workspace = true
 zksync_os_state_full_diffs.workspace = true
 zksync_os_merkle_tree.workspace = true
+versioned_merkle_tree.workspace = true
 zksync_os_raft.workspace = true
 zksync_os_tx_validators.workspace = true
 zksync_os_config_validation_macros.workspace = true
