+++ node/bin/src/prover_input_generator/tree_adapter.rs
@@ -1,4 +1,3 @@
-use super::PROVER_INPUT_GENERATOR_METRICS;
 use alloy::primitives::B256;
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::thread;
@@ -7,10 +6,12 @@ use zk_os_basic_system::system_implementation::flat_storage_model::FlatStorageLe
 use zk_os_basic_system_prev::system_implementation::flat_storage_model::FlatStorageLeaf as FlatStorageLeafDev;
 use zk_os_forward_system::run::{LeafProof, ReadStorage, ReadStorageTree};
 use zksync_os_batch_types::BlockMerkleTreeData;
-use zksync_os_merkle_tree::{
-    Blake2Hasher, HashTree, Leaf, MerkleTree, MerkleTreeProver, RocksDBWrapper, TreeOperation,
-    api::flat,
-};
+use zksync_os_merkle_tree::{Blake2Hasher, HashTree, Leaf, TreeOperation};
+
+// `VersionedMerkleTree` now lives in its own lib crate so the witness-assembly
+// code can reuse it without depending on `node/bin`. Re-export it here to keep
+// the existing `tree_adapter::VersionedMerkleTree` path valid for callers.
+pub(crate) use versioned_merkle_tree::VersionedMerkleTree;
 
 const TREE_DEPTH: u8 = 64;
 
@@ -154,154 +155,6 @@ impl TreeOutputAdapter {
         } else {
             Err(NoData)
         }
-    }
-}
-
-/// Storage adapter that reads data from the Merkle tree. This adapter is very inefficient in terms of I/O,
-/// but is universal as opposed to using a batch update proof (which will miss data for any keys
-/// not read / written in the batch).
-#[derive(Debug)]
-pub(super) struct VersionedMerkleTree {
-    inner: MerkleTree<RocksDBWrapper>,
-    version: u64,
-    cached_key_to_index: HashMap<B256, Option<u64>>,
-    cached_missing_key_to_prev_index: HashMap<B256, u64>,
-    cached_proofs: HashMap<u64, flat::StorageSlotProofEntryWithKey>,
-}
-
-impl VersionedMerkleTree {
-    pub(super) fn new(inner: MerkleTree<RocksDBWrapper>, version: u64) -> Self {
-        Self {
-            inner,
-            version,
-            cached_key_to_index: HashMap::new(),
-            cached_missing_key_to_prev_index: HashMap::new(),
-            cached_proofs: HashMap::new(),
-        }
-    }
-
-    fn read(&mut self, key: B256) -> Option<B256> {
-        let (proof, _) = self
-            .inner
-            .prove_flat(self.version, &[key])
-            .expect("failed getting Merkle proof")
-            .expect("tree version disappeared");
-        assert_eq!(
-            proof.len(),
-            1,
-            "sanity check failed: unexpected proof length"
-        );
-        let proof = proof.into_iter().next().unwrap();
-        let value = proof.value();
-
-        // Cache the proof since it's guaranteed to be requested later.
-        self.cache_proof(proof);
-
-        value
-    }
-
-    fn cache_proof(&mut self, proof: flat::StorageSlotProof) {
-        match proof.proof {
-            flat::InnerStorageSlotProof::Existing(entry) => {
-                self.insert_proof(proof.key, entry);
-            }
-            flat::InnerStorageSlotProof::NonExisting {
-                left_neighbor,
-                right_neighbor,
-            } => {
-                self.cached_key_to_index.insert(proof.key, None);
-                self.cached_missing_key_to_prev_index
-                    .insert(proof.key, left_neighbor.inner.index);
-                self.insert_proof(left_neighbor.leaf_key, left_neighbor.inner);
-                self.insert_proof(right_neighbor.leaf_key, right_neighbor.inner);
-            }
-        }
-    }
-
-    fn insert_proof(&mut self, key: B256, proof: flat::StorageSlotProofEntry) {
-        self.cached_key_to_index.insert(key, Some(proof.index));
-        self.cached_proofs.insert(
-            proof.index,
-            flat::StorageSlotProofEntryWithKey {
-                inner: proof,
-                leaf_key: key,
-            },
-        );
-    }
-
-    fn tree_index(&mut self, key: B256) -> Option<u64> {
-        if !self.cached_key_to_index.contains_key(&key) {
-            // Use proof API to get the necessary data. This is inefficient, but should (almost) never
-            // be triggered in practice.
-            self.read(key);
-        }
-        self.cached_key_to_index[&key]
-    }
-
-    fn merkle_proof(&mut self, tree_index: u64) -> LeafProof {
-        if !self.cached_proofs.contains_key(&tree_index) {
-            let proof = self
-                .inner
-                .prove_index_flat(self.version, tree_index)
-                .expect("failed getting Merkle proof")
-                .expect("tree version disappeared");
-            self.cached_proofs.insert(tree_index, proof);
-        }
-        Self::map_proof(&self.cached_proofs[&tree_index])
-    }
-
-    fn map_proof(proof: &flat::StorageSlotProofEntryWithKey) -> LeafProof {
-        let leaf = FlatStorageLeaf {
-            key: proof.leaf_key.0.into(),
-            value: proof.inner.value.0.into(),
-            next: proof.inner.next_index,
-        };
-
-        let mut merkle_path = Box::new([Bytes32::default(); 64]);
-        for (i, hash) in proof.inner.siblings.iter().enumerate() {
-            merkle_path[i] = hash.0.into();
-        }
-        // Fill in remaining Merkle path hashes from empty subtree hashes.
-        let merkle_path_len = proof.inner.siblings.len() as u8;
-        for level in merkle_path_len..TREE_DEPTH {
-            merkle_path[usize::from(level)] = Blake2Hasher.empty_subtree_hash(level).0.into();
-        }
-
-        LeafProof::new(proof.inner.index, leaf, merkle_path)
-    }
-
-    fn prev_tree_index(&mut self, key: B256) -> u64 {
-        if !self.cached_missing_key_to_prev_index.contains_key(&key) {
-            assert_eq!(self.read(key), None);
-        }
-        self.cached_missing_key_to_prev_index[&key]
-    }
-}
-
-/// Reports storage-related metrics on drop.
-impl Drop for VersionedMerkleTree {
-    fn drop(&mut self) {
-        if thread::panicking() {
-            return; // Do not report potentially incomplete data if generating prover input failed
-        }
-
-        PROVER_INPUT_GENERATOR_METRICS
-            .unexpected_queried_keys
-            .observe(self.cached_key_to_index.len());
-        PROVER_INPUT_GENERATOR_METRICS
-            .unexpected_queried_missing_keys
-            .observe(self.cached_missing_key_to_prev_index.len());
-        PROVER_INPUT_GENERATOR_METRICS
-            .unexpected_queried_proofs
-            .observe(self.cached_proofs.len());
-
-        tracing::info!(
-            version = self.version,
-            cached_key_to_index.len = self.cached_key_to_index.len(),
-            cached_missing_key_to_prev_index.len = self.cached_missing_key_to_prev_index.len(),
-            cached_proofs.len = self.cached_proofs.len(),
-            "finished providing storage via Merkle tree"
-        );
     }
 }
